requirement.txt
scrapy
json-log-formatter
boto3
sqlalchemy
psycopg2 
python-dotenv
apache-airflow[statsd]
dagster 
dagit 
pandera
sqlalchemy-redshift 
dbt-postgres
dbt-redshift
airflow-dbt
